pyproject.toml
[tool.poetry]
name = "testlikesoft"
version = "0.1.0"
description = ""
authors = ["samolin <samolin.vlad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
celery = "^5.3.6"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
